+++ src/contexts/auth.tsx
@@ -1,12 +1,12 @@
-import React, {createContext, useState, useEffect} from 'react';
-import auth, {FirebaseAuthTypes} from '@react-native-firebase/auth';
-import {useNavigation} from '@react-navigation/native';
-import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import {RootStackParamList} from 'src/App';
+import React, { createContext, useState, useEffect } from 'react';
+import auth, { FirebaseAuthTypes } from '@react-native-firebase/auth';
+import { useNavigation } from '@react-navigation/native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from 'src/App';
 import CreateUser from 'src/api/mutation/create-user.mutation.graphql';
 import CurrentUser from 'src/api/query/current-user.query.graphql';
-import {useMutation, ApolloError, useLazyQuery} from '@apollo/client';
-import {IUser} from 'src/utils/types/schema';
+import { useMutation, ApolloError, useLazyQuery } from '@apollo/client';
+import { IUser } from 'src/utils/types/schema';
 
 type NavigationProps = NativeStackScreenProps<RootStackParamList, 'Home'>;
 
@@ -50,7 +50,7 @@ export default function AuthContextProvider({
   const [error, setError] = useState<string | null>(null);
   const routeName = navigationState?.routes[navigationState.index].name;
 
-  const [createUser, {loading: apiLoading, error: apiError}] =
+  const [createUser, { loading: apiLoading, error: apiError }] =
     useMutation(CreateUser);
 
   const [
@@ -74,10 +74,10 @@ export default function AuthContextProvider({
           fireId: firebaseUser?.uid,
         },
       },
-      refetchQueries: [{query: CurrentUser}],
+      refetchQueries: [{ query: CurrentUser }],
       onCompleted: () => {
         auth()
-          .currentUser?.updateProfile({displayName: input.userName})
+          .currentUser?.updateProfile({ displayName: input.userName })
           .then(() => {
             navigation.navigate('Splash');
           });
